requirements.txt
# Python package dependencies
pandas>=1.3.5
numpy>=1.20.0
biopython>=1.79
pyyaml>=6.0
matplotlib>=3.5.1
seaborn>=0.11.2

# Core dependencies
python>=3.9
conda>=23.0.0  # For conda-based installation

# Metagenomics pipeline tools (handled by setup_tools.py)
# Quality Control
# - fastqc
# - multiqc

# Preprocessing
# - fastp
# - bwa
# - samtools

# Assembly
# - idba/idba_ud
# - idba_fq2fa
# - megahit
# - metaquast

# Binning
# - jgi_summarize_bam_contig_depths
# - metabat2
# - maxbin2
# - concoct
# - das_tool
# - metawrap

# Evaluation
# - checkm2
# - drep

# Taxonomy
# - gtdbtk
# - kraken2
# - bracken

# Annotation
# - eggnog-mapper/emapper.py
# - dbcan/run_dbcan
# - prodigal

# Note: This file lists dependencies for reference.
# Use setup_tools.py for installing the full pipeline environment